feature (RoomsController): #9 refactor RoomsController
