+++ BulgarskiChiflikGuestHouse.Web/GuestHouse.Web.csproj
@@ -18,8 +18,4 @@
     <ProjectReference Include="..\BulgarskiChiflikGuestHouse.Infrastructure\GuestHouse.Infrastructure.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Service\" />
-  </ItemGroup>
-
 </Project>
