+++ BulgarskiChiflikGuestHouse.Web/Service/IRoomWebService.cs
@@ -1,0 +1,10 @@
+﻿using GuestHouse.BLL.Service;
+using GuestHouse.Web.Models;
+
+namespace GuestHouse.Web.Service
+{
+	public interface IRoomWebService
+	{
+		Task<List<RoomViewModel>> GetAllRoomsAsync();
+	}
+}
