+++ BulgarskiChiflikGuestHouse.Web/Views/Rooms/GetAllRooms.cshtml
@@ -1,5 +1,4 @@
-﻿@using System.Diagnostics;
-@model List<RoomViewModel>
+﻿@model List<RoomViewModel>
 @{
     ViewData["Title"] = "Rooms";
 }
