+++ BulgarskiChiflikGuestHouse.Web/Views/Shared/_Layout.cshtml
@@ -64,7 +64,7 @@
                                 <a asp-controller="Home" asp-action="Index" class="nav-item nav-link active">Home</a>
                                 <a href="about.html" class="nav-item nav-link">About</a>
                                 <a href="service.html" class="nav-item nav-link">Menu</a>
-                                <a asp-controller="Rooms" asp-action="OurRooms" class="nav-item nav-link">Rooms</a>
+                                <a asp-controller="Rooms" asp-action="GetAllRooms" class="nav-item nav-link">Rooms</a>
                                 <a href="gallery.html" class="nav-item nav-link">Gallery</a>
                                 <a href="contact.html" class="nav-item nav-link">Contact</a>
                             </div>
